More code review feedback
