+++ cli/tests/subcommands/test_show.py
@@ -2,9 +2,10 @@ import logging
 import unittest
 import uuid
 
+import requests_mock
 from nose.plugins.attrib import attr
 
-from cook.subcommands.show import query_cluster
+from cook.subcommands.show import query_cluster, make_job_request
 
 
 @attr(cli=True)
@@ -14,11 +15,9 @@ class CookCliTest(unittest.TestCase):
     def setUp(self):
         self.logger = logging.getLogger(__name__)
 
-    def test_query_cluster(self):
-        cluster = {}
+    def test_query_cluster_should_gracefully_handle_json_parsing_failures(self):
+        cluster = {'url': 'http://localhost'}
         uuids = [uuid.uuid4()]
-
-        def make_job_request(_, __):
-            return {}
-
-        query_cluster(cluster, uuids, None, None, None, make_job_request, 'job')
+        with requests_mock.mock() as m:
+            m.get('http://localhost/rawscheduler', text='this is not json')
+            self.assertEqual({}, query_cluster(cluster, uuids, None, None, None, make_job_request, 'job'))
